pkg/cluster: if release image is passed always pass to installer
Currently if we explicitly send a release image we don't pass it to the installer unless it is a custom type.
But in the case of ci-op release images using --keep-bootstrap we are using the master installer. In which
case not passing the release image will result in using an unexpected release.

Signed-off-by: Sam Batschelet <sbatsche@redhat.com>